refactor(cli): split evaluate.rs and graphify.rs by concern (issue #104)
Both files passed max_function_complexity but failed the SIMPLE gate on
file-level cfg.cyclomatic — several modestly-complex functions bundled
into one file summed past the threshold even though none individually
exceeded it.

evaluate.rs mixed four concerns: GitNexus/MDG resolution, the shared
classify-with-representations step (inspect.rs already imported this
from evaluate — a sign it belonged to neither command), result
rendering, and run() orchestration. Split into composable.rs,
classify.rs, render.rs; evaluate.rs now holds only arg parsing + run().

graphify.rs bundled two independent subcommands (generate, orphans) in
one file. Split into a graphify/ submodule (generate.rs, orphans.rs),
leaving the parent file as arg-parsing + dispatch + the shared
print_json helper.

All existing tests moved with their code, unchanged.

Dogfooding with the v0.4.0 binary against topos-cli itself (issue #104):

  evaluate.rs: SLOP -> SIMPLE (cyclomatic 31->file split, was 22% simple)
  graphify.rs: SLOP -> SIMPLE (cyclomatic 20->file split, was 50% simple)

Directory rollup for topos/cli/src: SLOP -> SIMPLE + SECURE (14/14 files).
