+++ topos/cli/src/commands/evaluate.rs
@@ -30,19 +30,17 @@
 //! [`ModuleDependencyGraph`]: topos_engine::graphs::mdg::object::ModuleDependencyGraph
 //! [`ProgramDependenceGraph`]: topos_engine::graphs::pdg::object::ProgramDependenceGraph
 
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use clap::Args;
 use topos_engine::adapters::discovery::collect_source_files;
-use topos_engine::adapters::gitnexus::{current_git_branch, resolve_lbug_store};
-use topos_engine::core::characteristic_morphism::{CharacteristicMorphism, ClassificationResult};
+use topos_engine::core::characteristic_morphism::CharacteristicMorphism;
 use topos_engine::core::morphism::ProgramMorphism;
-use topos_engine::evaluation::policies::base::Priority;
-use topos_engine::functors::probes::uast::abstractness::AbstractnessRepresentation;
 use topos_engine::graphs::ast::languages::{language_file_suffixes, SUPPORTED_LANGUAGES};
-use topos_engine::graphs::base::Representation;
-use topos_engine::graphs::mdg::object::ModuleDependencyGraph;
-use topos_mcp::evaluation::ensure_gitnexus_dir;
+
+use super::classify::classify_with_representations;
+use super::composable::resolve_composable_mdg;
+use super::render::print_classification;
 
 #[derive(Args)]
 pub struct EvaluateArgs {
@@ -120,197 +118,4 @@ pub fn run(args: EvaluateArgs) -> Result<(), String> {
         }
     }
     Ok(())
-}
-
-/// Ensure a fresh `.gitnexus` build exists for `project_root` and load its
-/// `ModuleDependencyGraph`, or return `None` (with an explanatory `stderr`
-/// notice) if that isn't possible. Never returns an `Err` — COMPOSABLE is
-/// optional and its absence must not fail the whole evaluate run.
-///
-/// The resolve-or-generate decision itself (present/missing/stale, GitNexus
-/// availability, generation) is shared with the MCP evaluate tools via
-/// `topos_mcp::evaluation::ensure_gitnexus_dir` — this function only adds
-/// the CLI-specific "load once, reuse across every file in this run" MDG
-/// parsing on top (unlike MCP's `load_dep_graph`, which caches per
-/// `target_file` — a fit for arbitrary single-file tool calls, but N cache
-/// misses across a directory walk of N files).
-fn resolve_composable_mdg(
-    project_root: &Path,
-    gitnexus_dir_override: Option<&str>,
-) -> Option<ModuleDependencyGraph> {
-    let outcome = ensure_gitnexus_dir(
-        gitnexus_dir_override,
-        project_root,
-        /* skip = */ false,
-        /* capture = */ false,
-    );
-    if let Some(note) = &outcome.generation_note {
-        eprintln!("gitnexus: {note}");
-    }
-    let gitnexus_dir = outcome.gitnexus_dir?;
-
-    let branch = current_git_branch(project_root);
-    let resolved = resolve_lbug_store(&gitnexus_dir, branch.as_deref());
-    let lbug_path = match resolved.path {
-        Some(path) if path.is_dir() => path,
-        _ => {
-            eprintln!(
-                "gitnexus: no indexed store found at {} — evaluating SIMPLE/SECURE only.",
-                gitnexus_dir.display()
-            );
-            return None;
-        }
-    };
-
-    match ModuleDependencyGraph::from_json_dir(&lbug_path, project_root.to_string_lossy()) {
-        Ok(graph) => Some(graph),
-        Err(e) => {
-            eprintln!(
-                "gitnexus: failed to load dependency graph ({e}) — evaluating SIMPLE/SECURE only."
-            );
-            None
-        }
-    }
-}
-
-/// Build the CFG (SIMPLE), PDG (diagnostic), CPG (SECURE), and — when
-/// `mdg` is provided — MDG (COMPOSABLE) representations for `morphism`
-/// and classify it. `mdg` is expected to already have its `target_file`
-/// set to the file currently being classified.
-pub(crate) fn classify_with_representations(
-    classifier: &CharacteristicMorphism,
-    morphism: &mut ProgramMorphism,
-    mdg: Option<&ModuleDependencyGraph>,
-) -> ClassificationResult {
-    let cfg = morphism.build_cfg().cloned();
-    let pdg = morphism.build_pdg().cloned();
-    let cpg = morphism.build_cpg().cloned();
-    let abstractness = morphism
-        .ast
-        .as_ref()
-        .map(|ast| AbstractnessRepresentation::new(&ast.uast_root));
-    let mut representations: Vec<&dyn Representation> = Vec::new();
-    if let Some(cfg) = &cfg {
-        representations.push(cfg);
-    }
-    if let Some(pdg) = &pdg {
-        representations.push(pdg);
-    }
-    if let Some(cpg) = &cpg {
-        representations.push(cpg);
-    }
-    if let Some(abstractness) = &abstractness {
-        representations.push(abstractness);
-    }
-    if let Some(mdg) = mdg {
-        representations.push(mdg);
-    }
-    classifier.classify_detailed(morphism, &representations, Priority::default())
-}
-
-/// Print a verdict, per-generator scores, and raw metrics for one result.
-pub(crate) fn print_classification(result: &ClassificationResult) {
-    if !result.is_parseable {
-        println!("  {}", result.summary());
-        return;
-    }
-    println!("  Verdict: {}", result.summary());
-    for dim in ["simple", "composable", "secure"] {
-        let Some(val) = result.dimensions.get(dim) else {
-            continue;
-        };
-        let score = result.scores.get(dim).copied().unwrap_or(0.0) * 100.0;
-        println!("    {dim}: {val} [{score:.0}%]");
-    }
-    if !result.raw_metrics.is_empty() {
-        println!("  Raw metrics:");
-        let mut keys: Vec<&String> = result.raw_metrics.keys().collect();
-        keys.sort();
-        for key in keys {
-            let value = result.raw_metrics[key];
-            println!("    {key}: {value:.3}");
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-    use topos_engine::graphs::mdg::models::{GraphNode, GraphRelationship};
-
-    fn temp_dir(label: &str) -> PathBuf {
-        let dir = std::env::temp_dir().join(format!(
-            "topos_cli_evaluate_test_{label}_{}",
-            std::process::id()
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        dir
-    }
-
-    fn node(id: &str, label: &str, file_path: &str) -> GraphNode {
-        let mut properties = HashMap::new();
-        properties.insert(
-            "filePath".to_string(),
-            serde_json::Value::String(file_path.to_string()),
-        );
-        GraphNode {
-            id: id.to_string(),
-            label: label.to_string(),
-            properties,
-        }
-    }
-
-    fn rel(id: &str, source: &str, target: &str, rel_type: &str) -> GraphRelationship {
-        GraphRelationship {
-            id: id.to_string(),
-            source_id: source.to_string(),
-            target_id: target.to_string(),
-            rel_type: rel_type.to_string(),
-            confidence: 1.0,
-            reason: String::new(),
-            properties: HashMap::new(),
-        }
-    }
-
-    #[test]
-    fn classify_with_representations_scores_composable_when_mdg_present() {
-        let classifier = CharacteristicMorphism;
-        let mut morphism = ProgramMorphism::new("def f():\n    return 1\n", "python");
-
-        let mut mdg = ModuleDependencyGraph::new("a.py");
-        mdg.add_node(node("File:a.py", "File", "a.py"));
-        mdg.add_node(node("File:b.py", "File", "b.py"));
-        mdg.add_relationship(rel("i1", "File:a.py", "File:b.py", "IMPORTS"));
-
-        let without = classify_with_representations(&classifier, &mut morphism, None);
-        assert!(
-            !without.dimensions.contains_key("composable"),
-            "composable must not appear without an MDG representation"
-        );
-
-        let with = classify_with_representations(&classifier, &mut morphism, Some(&mdg));
-        assert!(
-            with.dimensions.contains_key("composable"),
-            "composable must appear once an MDG representation is attached"
-        );
-    }
-
-    #[test]
-    fn resolve_composable_mdg_returns_none_for_override_outside_project_root_without_shelling_out()
-    {
-        // An override outside `project_root` is rejected by
-        // `resolve_gitnexus_dir`/`depgraph_status` before any
-        // availability check or subprocess call, so this stays
-        // deterministic regardless of whether gitnexus happens to be
-        // installed on the machine running the test.
-        let project_root = temp_dir("root");
-        let outside = temp_dir("outside");
-
-        let result = resolve_composable_mdg(&project_root, Some(&outside.to_string_lossy()));
-        assert!(result.is_none());
-
-        std::fs::remove_dir_all(&project_root).ok();
-        std::fs::remove_dir_all(&outside).ok();
-    }
 }
